feat(slack,stt): extract shared STT package and add Slack voice message + status enhancements
- Extract platform-agnostic STT types (Transcriber, LocalSTT, PersistentSTT,
  FallbackSTT, AudioToPCM) from feishu/stt.go into new internal/messaging/stt/
- Add SlackConfig STT fields (STTProvider, STTLocalCmd, STTLocalMode, STTLocalIdleTTL)
- Wire Slack adapter STT with voice message classification and transcription
- Enhance Slack status text with tool input parameters and result previews
- Update AGENTS.md and STT-SETUP.md to reflect shared STT architecture
